chore: Update startup.sh and corpora.py
Update startup.sh and corpora.py files to ensure they are up to date with the latest changes.
