+++ .vscode/ltex.dictionary.en-US.txt
@@ -1,0 +1,1 @@
+Kaggle
