+++ informe/informe.out
@@ -1,0 +1,1 @@
+\BOOKMARK [0][-]{chapter.1}{\376\377\000I\000n\000f\000o\000r\000m\000e\000\040\000d\000e\000l\000\040\000p\000r\000o\000y\000e\000c\000t\000o\000\040\000d\000e\000\040\000s\000i\000s\000t\000e\000m\000a\000\040\000d\000e\000\040\000r\000e\000c\000u\000p\000e\000r\000a\000c\000i\000o\000n\000\040\000d\000e\000\040\000i\000n\000f\000o\000r\000m\000a\000c\000i\000\363\000n}{}% 1
