+++ startup.sh
